feat(Project): Forcing the whole project to be only a 32 bit project
Through the compiler argument -m32 of gcc/clang the compilation of the
project, for the emulation/tests, is now in 32 bit.

removing old hack for conditional arch compilation in 64 bit

+++ tests/linux_board/timer/timer.c
@@ -3,6 +3,7 @@
 #include <stdint.h>
 #include <unistd.h>
 #include <sys/time.h>
+#include <time.h>
 
 int8_t wait_milliseconds(const time_var_microseconds time)
 {
@@ -14,6 +15,6 @@ time_var_microseconds timer_time_now(void)
 {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return (time_var_microseconds)(tv.tv_sec SECONDS + tv.tv_usec); // Convert to microseconds
+    return (time_var_microseconds)( ( (uint64_t) tv.tv_sec) SECONDS + (uint64_t)tv.tv_usec); // Convert to microseconds
 
 }
